requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
defusedxml==0.7.1
Django==3.2.5
django-celery-beat==2.2.1
django-cors-headers==3.7.0
django-rest-framework==0.1.0
django-seed==0.2.2
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-xml==2.0.0
Faker==8.9.1
install==1.3.4
kombu==5.1.0
prompt-toolkit==3.0.19
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
vine==5.0.0
wcwidth==0.2.5
xmltodict==0.12.0
